Add allocation session to port allocator
Why:
The port pre-allocator closes all the ports it takes when it is
closed. So if you immediate create anothere pre-allocator after that
close and use it to allocate some ports, they could be the same as
the initial run. So we want to pre-allocate all the ports using the
same pre-allocator.